Fix issue with manifest conflicting with apps using library
The allowBackup and supportsRtl tags will conflict with applications that use this library but set a different value in the manifest for these settings, causeing build failures. While apps can configure their manifest to force override these settings, it is better just to disable them for library modules.

For more information on this issue, see https://bit.ly/2kkzXo0